feat(storybook) Structure Refactor
