+++ src/shared/shared.module.ts
@@ -2,7 +2,7 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
 // Stories
-import Button from '../stories/button.component';
+import Button from '../stories/button/button.component';
 
 const STORIES = [
   Button
