+++ src/stories/button/button.component.html
@@ -1,0 +1,7 @@
+<button
+    type="button"
+    (click)="onClick.emit($event)"
+    [ngClass]="classes"
+    [ngStyle]="{ 'background-color': backgroundColor }">
+    {{ label }}
+</button>
