+++ src/stories/button/button.component.ts
@@ -2,15 +2,8 @@ import { Component, Input, Output, EventEmitter } from '@angular/core';
 
 @Component({
   selector: 'storybook-button',
-  template: ` <button
-    type="button"
-    (click)="onClick.emit($event)"
-    [ngClass]="classes"
-    [ngStyle]="{ 'background-color': backgroundColor }"
-  >
-    {{ label }}
-  </button>`,
-  styleUrls: ['./button.css'],
+  templateUrl: './button.component.html',
+  styleUrls: ['./button.scss'],
 })
 export default class ButtonComponent {
   /**
