+++ src/stories/header/header.component.ts
@@ -1,0 +1,20 @@
+import { Component, Input, Output, EventEmitter } from '@angular/core';
+
+@Component({
+  selector: 'storybook-header',
+  templateUrl: './header.component.html',
+  styleUrls: ['./header.scss'],
+})
+export default class HeaderComponent {
+  @Input()
+  user: unknown = null;
+
+  @Output()
+  onLogin = new EventEmitter<Event>();
+
+  @Output()
+  onLogout = new EventEmitter<Event>();
+
+  @Output()
+  onCreateAccount = new EventEmitter<Event>();
+}
