+++ src/stories/header/header.stories.ts
@@ -3,7 +3,7 @@ import { CommonModule } from '@angular/common';
 // also exported from '@storybook/angular' if you can deal with breaking changes in 6.1
 import { Story, Meta } from '@storybook/angular/types-6-0';
 
-import Button from './button.component';
+import Button from '../button/button.component';
 import Header from './header.component';
 
 export default {
